[tool] Setup the Docker configuration
This commit provides to develop all services inside containers or
just use the database. The dev configuration checks if there is an
environment variable for the database URL. The containers are running
and share the ports with the host. It is possible to use the mix
phx.server at the host machine or at the container. That is the reason
why there are two different ports. Port 8080 for the host access the
container and 4000 to access the service up at the host (default).